requirements.txt
requests
python-dotenv
argparse
urllib3
datetime
socket
os
re
time